chore: [IAI-55] Remove cyclic dependencies (#3369)
* remove cyclic dependencies

* remove cie cyclic dependencies
