+++ ts/App.tsx
@@ -5,12 +5,10 @@ import { MenuProvider } from "react-native-popup-menu";
 import { Provider } from "react-redux";
 import { PersistGate } from "redux-persist/integration/react";
 
-import configureStoreAndPersistor from "./boot/configureStoreAndPersistor";
+import { store, persistor } from "./boot/configureStoreAndPersistor";
 import { LightModalProvider } from "./components/ui/LightModal";
 import RootContainer from "./RootContainer";
 import theme from "./theme";
-
-export const { store, persistor } = configureStoreAndPersistor();
 
 // Infer the `RootState` and `AppDispatch` types from the store itselfexport
 export type RootState = ReturnType<typeof store.getState>;
